@@ -1,10 +1,15 @@
-module github.com/tiaguinho/gosoap
+module github.com/SoMuchForSubtlety/gosoap
+
+go 1.21
 
 require (
-	github.com/google/go-cmp v0.5.0 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
-	golang.org/x/net v0.0.0-20210726213435-c6fcb2dbf985
-	gotest.tools v2.2.0+incompatible
+	github.com/stretchr/testify v1.8.4
+	golang.org/x/net v0.19.0
 )
 
-go 1.13
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	golang.org/x/text v0.14.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
