@@ -2,192 +2,203 @@ package gosoap
 
 import (
 	"bytes"
+	"context"
 	"encoding/xml"
 	"errors"
 	"fmt"
 	"io"
-	"io/ioutil"
+	"log/slog"
 	"net/http"
-	"net/http/httputil"
-	"net/url"
 	"strings"
-	"sync"
-	"time"
 
 	"golang.org/x/net/html/charset"
 )
 
-type SoapParams interface{}
-
-// HeaderParams holds params specific to the header
-type HeaderParams map[string]interface{}
-
 // Params type is used to set the params in soap request
-type Params map[string]interface{}
-type ArrayParams [][2]interface{}
-type SliceParams []interface{}
+type Params map[string]any
 
-type DumpLogger interface {
-	LogRequest(method string, dump []byte)
-	LogResponse(method string, dump []byte)
-}
-
-type fmtLogger struct{}
-
-func (l *fmtLogger) LogRequest(method string, dump []byte) {
-	fmt.Printf("Request:\n%v\n----\n", string(dump))
-}
-
-func (l *fmtLogger) LogResponse(method string, dump []byte) {
-	fmt.Printf("Response:\n%v\n----\n", string(dump))
-}
+type (
+	ArrayParams [][2]any
+	SliceParams []any
+)
 
 // Config config the Client
 type Config struct {
-	Dump   bool
-	Logger DumpLogger
+	Client      *http.Client
+	AutoAction  bool
+	LogRequests bool
+	Logger      CommunicationLogger
+
+	Service string
+	Port    string
+
+	EnvelopePrefix string
+	EnvelopeAttrs  map[string]string
+
+	Username string
+	Password string
 }
 
-// SoapClient return new *Client to handle the requests with the WSDL
-func SoapClient(wsdl string, httpClient *http.Client) (*Client, error) {
-	return SoapClientWithConfig(wsdl, httpClient, &Config{Dump: false, Logger: &fmtLogger{}})
-}
+// NewClient return new *Client to handle the requests with the WSDL
+func NewClient(wsdlSource WSDLSource, config *Config) (*Client, error) {
+	if config == nil {
+		config = &Config{}
+	}
 
-// SoapClientWithConfig return new *Client to handle the requests with the WSDL
-func SoapClientWithConfig(wsdl string, httpClient *http.Client, config *Config) (*Client, error) {
-	_, err := url.Parse(wsdl)
+	if config.Client == nil {
+		config.Client = http.DefaultClient
+	}
+
+	if config.Logger == nil {
+		// default to info because the default logger has debug logs disabled
+		config.Logger = NewSlogAdapter(slog.Default(), slog.LevelInfo)
+	}
+
+	if config.EnvelopePrefix == "" {
+		config.EnvelopePrefix = "soap"
+	}
+	if len(config.EnvelopeAttrs) == 0 {
+		config.EnvelopeAttrs = map[string]string{
+			"xmlns:xsi":  "http://www.w3.org/2001/XMLSchema-instance",
+			"xmlns:xsd":  "http://www.w3.org/2001/XMLSchema",
+			"xmlns:soap": "http://schemas.xmlsoap.org/soap/envelope/",
+		}
+	}
+
+	definitions, err := getWSDLDefinitions(wsdlSource, config)
 	if err != nil {
 		return nil, err
 	}
 
-	if httpClient == nil {
-		httpClient = &http.Client{}
+	var namespace string
+	if definitions.Types != nil {
+		// FIXME: this can be incorrect, we need to add a config option to select a type element (by targetNamespace?)
+		schema := definitions.Types[0].XsdSchema[0]
+		namespace = schema.TargetNamespace
+		if namespace == "" && len(schema.Imports) > 0 {
+			namespace = schema.Imports[0].Namespace
+		}
 	}
 
-	if config.Logger == nil {
-		config.Logger = &fmtLogger{}
+	service, port, err := definitions.serviceAndPort(config.Service, config.Port)
+	if err != nil {
+		return nil, fmt.Errorf("could not determin SOAP address: %w", err)
 	}
+	config.Service = service.Name
+	config.Port = port.Name
 
-	c := &Client{
-		wsdl:       wsdl,
-		config:     config,
-		HTTPClient: httpClient,
-		AutoAction: false,
+	bindingName := port.Binding
+	splitBindingName := strings.Split(port.Binding, ":")
+	if len(splitBindingName) == 2 {
+		bindingName = splitBindingName[1] // strip off the namespace
 	}
+	var binding *wsdlBinding
+	for _, b := range definitions.Bindings {
+		if b.Name == bindingName || b.Name == port.Binding {
+			binding = b
+			break
+		}
+	}
+	if binding == nil {
+		return nil, fmt.Errorf("could not find binding matching %q", port.Binding)
+	}
+	return &Client{
+		config:        *config,
+		httpClient:    config.Client,
+		binding:       binding,
+		autoActionURL: strings.TrimSuffix(definitions.TargetNamespace, "/"),
+		address:       port.SoapAddresses[0].Location, // TODO: use multiple addresses?
+		namespace:     namespace,
+	}, nil
+}
 
-	return c, nil
+func (d *wsdlDefinitions) serviceAndPort(serviceName, portName string) (*wsdlService, *wsdlPort, error) {
+	var service *wsdlService
+	if len(d.Services) == 0 {
+		return nil, nil, errors.New("WSDL has no services")
+	}
+	if serviceName == "" {
+		service = d.Services[0]
+	} else {
+		var possibleServices []string
+		for _, svc := range d.Services {
+			possibleServices = append(possibleServices, svc.Name)
+			if svc.Name == serviceName {
+				service = svc
+				break
+			}
+		}
+		if service == nil {
+			return nil, nil, fmt.Errorf("no service matching %q found, possible values are %v", serviceName, possibleServices)
+		}
+	}
+	if len(service.Ports) == 0 {
+		return nil, nil, fmt.Errorf("WSDL service %q has no ports", service.Name)
+	}
+	var port *wsdlPort
+	if portName == "" {
+		port = service.Ports[0]
+	} else {
+		var possiblePorts []string
+		for _, p := range service.Ports {
+			possiblePorts = append(possiblePorts, p.Name)
+			if p.Name == portName {
+				port = p
+				break
+			}
+		}
+		if port == nil {
+			return nil, nil, fmt.Errorf("no port matching %q found, possible values are %v", portName, possiblePorts)
+		}
+	}
+	if len(port.SoapAddresses) == 0 {
+		return nil, nil, fmt.Errorf("WSDL port %q has no addresses", port.Name)
+	}
+	return service, port, nil
 }
 
 // Client struct hold all the information about WSDL,
 // request and response of the server
 type Client struct {
-	HTTPClient   *http.Client
-	AutoAction   bool
-	URL          string
-	HeaderName   string
-	HeaderParams SoapParams
-	Definitions  *wsdlDefinitions
-	// Must be set before first request otherwise has no effect, minimum is 15 minutes.
-	RefreshDefinitionsAfter time.Duration
-	Username                string
-	Password                string
+	httpClient *http.Client
+	config     Config
 
-	once                 sync.Once
-	definitionsErr       error
-	onRequest            sync.WaitGroup
-	onDefinitionsRefresh sync.WaitGroup
-	wsdl                 string
-	config               *Config
+	address       string
+	namespace     string
+	autoActionURL string
+	binding       *wsdlBinding
 }
 
-// Call call's the method m with Params p
-func (c *Client) Call(m string, p SoapParams) (res *Response, err error) {
-	return c.Do(NewRequest(m, p))
-}
-
-// CallByStruct call's by struct
-func (c *Client) CallByStruct(s RequestStruct) (res *Response, err error) {
-	req, err := NewRequestByStruct(s)
-	if err != nil {
-		return nil, err
-	}
-
-	return c.Do(req)
-}
-
-func (c *Client) waitAndRefreshDefinitions(d time.Duration) {
-	for {
-		time.Sleep(d)
-		c.onRequest.Wait()
-		c.onDefinitionsRefresh.Add(1)
-		c.initWsdl()
-		c.onDefinitionsRefresh.Done()
-	}
-}
-
-func (c *Client) initWsdl() {
-	c.Definitions, c.definitionsErr = getWsdlDefinitions(c.wsdl, c.HTTPClient)
-	if c.definitionsErr == nil {
-		c.URL = strings.TrimSuffix(c.Definitions.TargetNamespace, "/")
-	}
-}
-
-// SetWSDL set WSDL url
-func (c *Client) SetWSDL(wsdl string) {
-	c.onRequest.Wait()
-	c.onDefinitionsRefresh.Wait()
-	c.onRequest.Add(1)
-	c.onDefinitionsRefresh.Add(1)
-	defer c.onRequest.Done()
-	defer c.onDefinitionsRefresh.Done()
-	c.wsdl = wsdl
-	c.initWsdl()
+func (c *Client) Call(ctx context.Context, wsdlOperation string, body any, headerParams ...any) (res *Response, err error) {
+	return c.Do(ctx, NewRequest(wsdlOperation, body, headerParams...))
 }
 
 // Do Process Soap Request
-func (c *Client) Do(req *Request) (res *Response, err error) {
-	c.onDefinitionsRefresh.Wait()
-	c.onRequest.Add(1)
-	defer c.onRequest.Done()
-
-	c.once.Do(func() {
-		c.initWsdl()
-		// 15 minute to prevent abuse.
-		if c.RefreshDefinitionsAfter >= 15*time.Minute {
-			go c.waitAndRefreshDefinitions(c.RefreshDefinitionsAfter)
+func (c *Client) Do(ctx context.Context, req *Request) (res *Response, err error) {
+	var action string
+	if c.config.AutoAction {
+		action = fmt.Sprintf("%s/%s/%s", c.autoActionURL, c.config.Service, req.WSDLOperation)
+	} else {
+		action, err = c.binding.GetSoapActionFromWsdlOperation(req.WSDLOperation)
+		if err != nil {
+			return nil, err
 		}
-	})
-
-	if c.definitionsErr != nil {
-		return nil, c.definitionsErr
 	}
-
-	if c.Definitions == nil {
-		return nil, errors.New("wsdl definitions not found")
-	}
-
-	if c.Definitions.Services == nil {
-		return nil, errors.New("No Services found in wsdl definitions")
-	}
-
 	p := &process{
-		Client:     c,
-		Request:    req,
-		SoapAction: c.Definitions.GetSoapActionFromWsdlOperation(req.Method),
+		config:     &c.config,
+		namespace:  c.namespace,
+		request:    req,
+		soapAction: action,
 	}
 
-	if p.SoapAction == "" && c.AutoAction {
-		p.SoapAction = fmt.Sprintf("%s/%s/%s", c.URL, c.Definitions.Services[0].Name, req.Method)
-	}
-
-	p.Payload, err = xml.MarshalIndent(p, "", "    ")
+	p.payload, err = xml.MarshalIndent(p, "", "    ")
 	if err != nil {
 		return nil, err
 	}
 
-	b, err := p.doRequest(c.Definitions.Services[0].Ports[0].SoapAddresses[0].Location)
+	b, err := c.doRequest(ctx, p)
 	if err != nil {
-		return nil, ErrorWithPayload{err, p.Payload}
+		return nil, ErrorWithPayload{err, p.payload}
 	}
 
 	var soap SoapEnvelope
@@ -201,84 +212,86 @@ func (c *Client) Do(req *Request) (res *Response, err error) {
 	err = decoder.Decode(&soap)
 
 	res = &Response{
-		Body:    soap.Body.Contents,
-		Header:  soap.Header.Contents,
-		Payload: p.Payload,
+		Body:          soap.Body.Contents,
+		HeaderEntries: soap.Header.Contents,
 	}
 	if err != nil {
-		return res, ErrorWithPayload{err, p.Payload}
+		return res, ErrorWithPayload{err, p.payload}
 	}
 
 	return res, nil
 }
 
 type process struct {
-	Client     *Client
-	Request    *Request
-	SoapAction string
-	Payload    []byte
+	config    *Config
+	request   *Request
+	namespace string
+	// see https://www.w3.org/TR/2000/NOTE-SOAP-20000508/#_Toc478383528
+	soapAction string
+	payload    []byte
 }
 
 // doRequest makes new request to the server using the c.Method, c.URL and the body.
 // body is enveloped in Do method
-func (p *process) doRequest(url string) ([]byte, error) {
-	req, err := http.NewRequest("POST", url, bytes.NewBuffer(p.Payload))
+func (c *Client) doRequest(ctx context.Context, p *process) ([]byte, error) {
+	req, err := http.NewRequestWithContext(ctx, http.MethodPost, c.address, bytes.NewBuffer(p.payload))
 	if err != nil {
 		return nil, err
 	}
 
-	if p.Client.config != nil && p.Client.config.Dump {
-		dump, err := httputil.DumpRequestOut(req, true)
+	if c.config.LogRequests {
+		var body []byte
+		req.Body, body, err = drainBody(req.Body)
 		if err != nil {
 			return nil, err
 		}
-		p.Client.config.Logger.LogRequest(p.Request.Method, dump)
+		c.config.Logger.LogRequest(p.request.WSDLOperation, req.Header, body)
 	}
 
-	if p.Client.Username != "" && p.Client.Password != "" {
-		req.SetBasicAuth(p.Client.Username, p.Client.Password)
+	if c.config.Username != "" && c.config.Password != "" {
+		req.SetBasicAuth(c.config.Username, c.config.Password)
 	}
 
-	req.ContentLength = int64(len(p.Payload))
+	req.ContentLength = int64(len(p.payload))
 
 	req.Header.Add("Content-Type", "text/xml;charset=UTF-8")
 	req.Header.Add("Accept", "text/xml")
-	if p.SoapAction != "" {
-		req.Header.Add("SOAPAction", p.SoapAction)
+	if p.soapAction != "" {
+		req.Header.Add("SOAPAction", p.soapAction)
 	}
 
-	resp, err := p.httpClient().Do(req)
+	resp, err := c.httpClient.Do(req)
 	if err != nil {
 		return nil, err
 	}
 	defer resp.Body.Close()
 
-	if p.Client.config != nil && p.Client.config.Dump {
-		dump, err := httputil.DumpResponse(resp, true)
+	if c.config.LogRequests {
+		var body []byte
+		resp.Body, body, err = drainBody(resp.Body)
 		if err != nil {
 			return nil, err
 		}
-		p.Client.config.Logger.LogResponse(p.Request.Method, dump)
+		c.config.Logger.LogResponse(p.request.WSDLOperation, req.Header, body)
 	}
 
-	if resp.StatusCode < 200 || resp.StatusCode >= 400 {
-		if !(p.Client.config != nil && p.Client.config.Dump) {
-			_, err := io.Copy(ioutil.Discard, resp.Body)
-			if err != nil {
-				return nil, err
-			}
-		}
-		return nil, errors.New("unexpected status code: " + resp.Status)
-	}
-
-	return ioutil.ReadAll(resp.Body)
+	return io.ReadAll(resp.Body)
 }
 
-func (p *process) httpClient() *http.Client {
-	if p.Client.HTTPClient != nil {
-		return p.Client.HTTPClient
+// from net/http/httputil
+func drainBody(b io.ReadCloser) (r1 io.ReadCloser, r2 []byte, err error) {
+	if b == nil || b == http.NoBody {
+		// No copying needed. Preserve the magic sentinel meaning of NoBody.
+		return http.NoBody, nil, nil
 	}
-	return http.DefaultClient
+	var buf bytes.Buffer
+	if _, err = buf.ReadFrom(b); err != nil {
+		return nil, nil, err
+	}
+	if err = b.Close(); err != nil {
+		return nil, nil, err
+	}
+	return io.NopCloser(&buf), buf.Bytes(), nil
 }
 
 // ErrorWithPayload error payload schema
