@@ -1,77 +1,80 @@
 package gosoap
 
 import (
+	"context"
 	"crypto/tls"
-	"fmt"
-	"log"
 	"net/http"
-	"regexp"
 	"testing"
+	"time"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
-var (
-	scts = []struct {
-		URL    string
-		Err    bool
-		Client *http.Client
-	}{
-		{
-			URL: "://www.server",
-			Err: false,
-		},
-		{
-			URL: "",
-			Err: false,
-		},
-		{
-			URL: "http://ec.europa.eu/taxation_customs/vies/checkVatService.wsdl",
-			Err: true,
-		},
-		{
-			URL: "http://ec.europa.eu/taxation_customs/vies/checkVatService.wsdl",
-			Err: true,
-			Client: &http.Client{
-				Transport: &http.Transport{
-					TLSClientConfig: &tls.Config{
-						InsecureSkipVerify: true,
-					},
+var scts = []struct {
+	name        string
+	source      WSDLSource
+	expectedErr string
+	client      *http.Client
+}{
+	{
+		name:        "invalid URL",
+		source:      SourceFromURI("://www.server"),
+		expectedErr: "could not load WSDL: could not fetch WSDL resource: parse \"://www.server\": missing protocol scheme",
+	},
+	{
+		name:        "no URL",
+		source:      SourceFromURI(""),
+		expectedErr: "could not load WSDL: could not fetch WSDL resource: Get \"\": unsupported protocol scheme \"\"",
+	},
+	{
+		name:   "good URL",
+		source: vatSource,
+	},
+	{
+		name:   "custom client",
+		source: vatSource,
+		client: &http.Client{
+			Transport: &http.Transport{
+				TLSClientConfig: &tls.Config{
+					InsecureSkipVerify: true,
 				},
 			},
 		},
-	}
-)
+	},
+}
 
 func TestSoapClient(t *testing.T) {
+	t.Parallel()
 	for _, sct := range scts {
-		_, err := SoapClient(sct.URL, nil)
-		if err != nil && sct.Err {
-			t.Errorf("URL: %s - error: %s", sct.URL, err)
-		}
+		sct := sct
+		t.Run(sct.name, func(t *testing.T) {
+			t.Parallel()
+			_, err := NewClient(sct.source, nil)
+			if sct.expectedErr != "" {
+				assert.EqualError(t, err, sct.expectedErr)
+			} else {
+				assert.NoError(t, err)
+			}
+		})
 	}
 }
 
-func TestSoapClienWithClient(t *testing.T) {
-	client, err := SoapClient(scts[3].URL, scts[3].Client)
+func TestSoapClientWithClient(t *testing.T) {
+	t.Parallel()
+	client, err := NewClient(scts[3].source, &Config{
+		Client: scts[3].client,
+	})
 
-	if client.HTTPClient != scts[3].Client {
+	if client.httpClient != scts[3].client {
 		t.Errorf("HTTP client is not the same as in initialization: - error: %s", err)
 	}
-
-	if err != nil {
-		t.Errorf("URL: %s - error: %s", scts[3].URL, err)
-	}
+	assert.NoError(t, err)
 }
 
 type CheckVatRequest struct {
-	CountryCode string
-	VatNumber   string
-}
-
-func (r CheckVatRequest) SoapBuildRequest() *Request {
-	return NewRequest("checkVat", Params{
-		"countryCode": r.CountryCode,
-		"vatNumber":   r.VatNumber,
-	})
+	CountryCode string `xml:"countryCode"`
+	VatNumber   string `xml:"vatNumber"`
 }
 
 type CheckVatResponse struct {
@@ -91,222 +94,87 @@ type NumberToWordsResponse struct {
 	NumberToWordsResult string
 }
 
-type WhoisResponse struct {
-	WhoisResult string
-}
-
-var (
-	rv CheckVatResponse
-	rc CapitalCityResponse
-	rn NumberToWordsResponse
-	rw WhoisResponse
-
-	params = Params{}
-)
-
-func TestClient_Call(t *testing.T) {
-	soap, err := SoapClientWithConfig("http://ec.europa.eu/taxation_customs/vies/checkVatService.wsdl",
-		nil,
-		&Config{Dump: true},
-	)
-	if err != nil {
-		t.Errorf("error not expected: %s", err)
-	}
-
-	var res *Response
-
-	params["vatNumber"] = "6388047V"
-	params["countryCode"] = "IE"
-	res, err = soap.Call("", params)
-	if err == nil {
-		t.Errorf("method is empty")
-	}
-
-	if res != nil {
-		t.Errorf("body is empty")
-	}
-
-	res, err = soap.Call("checkVat", params)
-	if err != nil {
-		t.Errorf("error in soap call: %s", err)
-	}
-
-	res.Unmarshal(&rv)
-	if rv.CountryCode != "IE" {
-		t.Errorf("error: %+v", rv)
-	}
-
-	soap, err = SoapClient("http://webservices.oorsprong.org/websamples.countryinfo/CountryInfoService.wso?WSDL", nil)
-	if err != nil {
-		t.Errorf("error not expected: %s", err)
-	}
-
-	res, err = soap.Call("CapitalCity", Params{"sCountryISOCode": "GB"})
-	if err != nil {
-		t.Errorf("error in soap call: %s", err)
-	}
-
-	res.Unmarshal(&rc)
-
-	if rc.CapitalCityResult != "London" {
-		t.Errorf("error: %+v", rc)
-	}
-
-	soap, err = SoapClient("http://www.dataaccess.com/webservicesserver/numberconversion.wso?WSDL", nil)
-	if err != nil {
-		t.Errorf("error not expected: %s", err)
-	}
-
-	res, err = soap.Call("NumberToWords", Params{"ubiNum": "23"})
-	if err != nil {
-		t.Errorf("error in soap call: %s", err)
-	}
-
-	res.Unmarshal(&rn)
-
-	if rn.NumberToWordsResult != "twenty three " {
-		t.Errorf("error: %+v", rn)
-	}
-
-	soap, err = SoapClient("https://domains.livedns.co.il/API/DomainsAPI.asmx?WSDL", nil)
-	if err != nil {
-		t.Errorf("error not expected: %s", err)
-	}
-
-	res, err = soap.Call("Whois", Params{"DomainName": "google.com"})
-	if err != nil {
-		t.Errorf("error in soap call: %s", err)
-	}
-
-	res.Unmarshal(&rw)
-
-	if rw.WhoisResult != "0" {
-		t.Errorf("error: %+v", rw)
-	}
-
-	c := &Client{}
-	res, err = c.Call("", Params{})
-	if err == nil {
-		t.Errorf("error expected but nothing got.")
-	}
-
-	c.SetWSDL("://test.")
-
-	res, err = c.Call("checkVat", params)
-	if err == nil {
-		t.Errorf("invalid WSDL")
-	}
-}
-
-type customLogger struct{}
-
-func (c customLogger) LogRequest(method string, dump []byte) {
-	var re = regexp.MustCompile(`(<vatNumber>)[\s\S]*?(<\/vatNumber>)`)
-	maskedResponse := re.ReplaceAllString(string(dump), `${1}XXX${2}`)
-
-	log.Println(fmt.Sprintf("%s request: %s", method, maskedResponse))
-}
-
-func (c customLogger) LogResponse(method string, dump []byte) {
-	if method == "checkVat" {
-		return
-	}
-
-	log.Println(fmt.Sprintf("Response: %s", dump))
-}
-
-func TestClient_Call_WithCustomLogger(t *testing.T) {
-	soap, err := SoapClientWithConfig("http://ec.europa.eu/taxation_customs/vies/checkVatService.wsdl",
-		nil,
-		&Config{Dump: true, Logger: &customLogger{}},
-	)
-	if err != nil {
-		t.Errorf("error not expected: %s", err)
-	}
-
-	var res *Response
-
-	res, err = soap.CallByStruct(CheckVatRequest{
-		CountryCode: "IE",
-		VatNumber:   "6388047V",
-	})
-	if err != nil {
-		t.Errorf("error in soap call: %s", err)
-	}
-
-	res.Unmarshal(&rv)
-	if rv.CountryCode != "IE" {
-		t.Errorf("error: %+v", rv)
-	}
-
-	_, err = soap.CallByStruct(nil)
-	if err == nil {
-		t.Error("err can't be nil")
-	}
-}
-
-func TestClient_CallByStruct(t *testing.T) {
-	soap, err := SoapClient("http://ec.europa.eu/taxation_customs/vies/checkVatService.wsdl", nil)
-	if err != nil {
-		t.Errorf("error not expected: %s", err)
-	}
-
-	var res *Response
-
-	res, err = soap.CallByStruct(CheckVatRequest{
-		CountryCode: "IE",
-		VatNumber:   "6388047V",
-	})
-	if err != nil {
-		t.Errorf("error in soap call: %s", err)
-	}
-
-	res.Unmarshal(&rv)
-	if rv.CountryCode != "IE" {
-		t.Errorf("error: %+v", rv)
-	}
-
-	_, err = soap.CallByStruct(nil)
-	if err == nil {
-		t.Error("err can't be nil")
-	}
-}
-
-func TestClient_Call_NonUtf8(t *testing.T) {
-	soap, err := SoapClient("https://demo.ilias.de/webservice/soap/server.php?wsdl", nil)
-	if err != nil {
-		t.Errorf("error not expected: %s", err)
-	}
-
-	_, err = soap.Call("login", Params{"client": "demo", "username": "robert", "password": "iliasdemo"})
-	if err != nil {
-		t.Errorf("error in soap call: %s", err)
-	}
-}
-
-func TestProcess_doRequest(t *testing.T) {
-	c := &process{
-		Client: &Client{
-			HTTPClient: &http.Client{},
+func TestValidRequests(t *testing.T) {
+	t.Parallel()
+	loc, err := time.LoadLocation("CET")
+	require.NoError(t, err)
+	testCases := []struct {
+		name         string
+		wsdl         string
+		operation    string
+		request      any
+		resp         any
+		expectedResp any
+	}{
+		{
+			name:      "vat",
+			wsdl:      "https://ec.europa.eu/taxation_customs/vies/checkVatService.wsdl",
+			operation: "checkVat",
+			request: Params{
+				"vatNumber":   "6388047V",
+				"countryCode": "IE",
+			},
+			resp: &CheckVatResponse{},
+			expectedResp: &CheckVatResponse{
+				CountryCode: "IE",
+				VatNumber:   "6388047V",
+				Name:        "GOOGLE IRELAND LIMITED",
+				Address:     "3RD FLOOR, GORDON HOUSE, BARROW STREET, DUBLIN 4",
+				Valid:       "true",
+				RequestDate: time.Now().In(loc).Format("2006-01-02-07:00"),
+			},
+		},
+		{
+			name:      "capital",
+			wsdl:      "http://webservices.oorsprong.org/websamples.countryinfo/CountryInfoService.wso?WSDL",
+			operation: "CapitalCity",
+			request:   Params{"sCountryISOCode": "GB"},
+			resp:      &CapitalCityResponse{},
+			expectedResp: &CapitalCityResponse{
+				CapitalCityResult: "London",
+			},
+		},
+		{
+			name:      "numbers",
+			wsdl:      "https://www.dataaccess.com/webservicesserver/numberconversion.wso?WSDL",
+			operation: "NumberToWords",
+			request:   Params{"ubiNum": "23"},
+			resp:      &NumberToWordsResponse{},
+			expectedResp: &NumberToWordsResponse{
+				NumberToWordsResult: "twenty three ",
+			},
 		},
 	}
 
-	_, err := c.doRequest("")
-	if err == nil {
-		t.Errorf("body is empty")
+	for _, tc := range testCases {
+		tc := tc
+		t.Run(tc.name, func(t *testing.T) {
+			t.Parallel()
+			client, err := NewClient(SourceFromURI(tc.wsdl), &Config{LogRequests: true})
+			require.NoError(t, err)
+			res, err := client.Call(context.Background(), tc.operation, tc.request)
+			require.NoError(t, err)
+			assert.NotNil(t, res)
+			err = res.Unmarshal(tc.resp)
+			assert.NoError(t, err)
+			assert.Equal(t, tc.expectedResp, tc.resp)
+		})
 	}
+}
 
-	_, err = c.doRequest("://teste.")
-	if err == nil {
-		t.Errorf("invalid WSDL")
-	}
+func TestInvalidWSDL(t *testing.T) {
+	t.Parallel()
+	c := &Client{}
+	_, err := c.Call(context.Background(), "", Params{})
+	assert.ErrorContains(t, err, "WSDL binding is nil")
+}
 
-	_, err = c.doRequest("https://google.com/non-existent-url")
-	if err == nil {
-		t.Errorf("err can't be nil")
-	}
+func TestClient_Call_NonUtf8(t *testing.T) {
+	t.Skip("server is down")
+	t.Parallel()
+	soap, err := NewClient(SourceFromURI("https://demo.ilias.de/webservice/soap/server.php?wsdl"), nil)
+	assert.NoError(t, err)
 
-	if err != nil && err.Error() != "unexpected status code: 404 Not Found" {
-		t.Errorf("unexpected error: %s", err)
-	}
+	_, err = soap.Call(context.Background(), "login", Params{"client": "demo", "username": "robert", "password": "iliasdemo"})
+	assert.NoError(t, err)
 }
